renamed files and moved to the general utils
